backend/requirements.txt
Django==5.1.6
djangorestframework==3.17.1
drf-spectacular==0.29.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
gunicorn==23.0.0
djangorestframework-simplejwt==5.3.1
redis==6.4.0
django-redis==5.4.0